try catch no user.service

+++ ghiblix-api/src/modules/users/users.service.ts
@@ -1,5 +1,4 @@
-import { Inject, Injectable } from '@nestjs/common';
-import { exception } from 'console';
+import { Inject, Injectable, NotFoundException, NotAcceptableException } from '@nestjs/common';
 import { User } from './user.entity';
 
 @Injectable()
@@ -21,27 +20,53 @@ export class UsersService {
         return await this.usersRepository.findOne<User>({where: { email } });
     }
     
-    create(user: User): Promise<User> {
+    async create(user: User): Promise<User> {
+
+        try {
+
+            const email = user.email;
+            const exists = await this.usersRepository.findOne({where: {email: email}});
+
+            if(exists)
+                throw new NotAcceptableException();
+
+        } catch (error) {
+            throw error;
+        }
         return this.usersRepository.create(user);
     }
 
     async update(id:number, user: User): Promise<User> {
 
         try {
-            if(!this.usersRepository.findByPk(id))
-                throw 'ID '
+
+            const exists = await this.usersRepository.findOne({where: {id: id}});
+
+            if(!exists)
+                throw new NotFoundException();
 
             else {
-                await this.usersRepository.update(user , { where: {id: id} });
+                await this.usersRepository.update(user, { where: {id: id} });
                 return this.getById(id);
             }
             
-        } catch(err) {
-
+        } catch(error) {
+            throw error;                
         }
     }
 
     async delete(id: number) {
-        this.usersRepository.destroy({where: {id: id}});
+        try {
+            const exists = await this.usersRepository.findOne({where: {id: id}});
+
+            if(!exists)
+                throw new NotFoundException();
+            else 
+                this.usersRepository.destroy({where: {id: id}});
+
+        } catch (error) {
+            throw error;
+        }
+        
     }
 }
